Add oberon::renderer_3d.
This moves some aspects of the oberon::window object (those related to
the Vulkan swapchain) into a separate oberon::renderer_3d object.
oberon::renderer_3d further manages Vulkan render passes and
framebuffers. The intent is for renderer_3d to also manage compiled
Vulkan pipelines in the future.

This commit also removes the oberon::child_object type in favor of
adding its functionality directly to oberon::object. This reduces a lot
of boilerplate related to walking up a chain of parents.

e.g. retrieving a context from the parent (a window) of a renderer_3d.

As a result it is potentially possible to invoke
oberon::object::parent() on an object with no parent (i.e. one where the
parent pointer is null). This is bad. Don't do it.

I've also removed all of the d_ptr/q_ptr stuff from oberon::object. This
strategy sort of made sense to me when I was just copying Qt. I have
replaced it with a templated oberon::reference_cast function in
types.hpp. This significantly reduces the amount of typing I need to do
to cast implementation references and parent references to their correct
types.