Improved object dependency handling.
Adds some changes to dependencies so that objects with 2 or more
dependencies are easier to handle. Also adds some giant macros to make
it relatively easy to add simple code to a class for resolving or
storing dependencies.

Additionally, there are some fixes to oberon/debug.hpp related to _MSG
macro variants. These should have been obvious to me earlier but oh well.