Refocus io_subsystem on Linux target.
As part of an effort to refocus on more specific goals for oberon (e.g.,
Linux with X11 as a target system) this reorients the
oberon::io_subsystem object specifically towards an X11 connected
system. It also reduces some of the overhead of things like loading EWMH
atoms because I understand X11 a little better now.