Rename environment to platform.
Renames the environment object to oberon::platform. This seems more
consistent with my intent.

Adds arguments (i.e., argc and argv) to oberon::application::run. These
can be used during system initialization (as they are in the linux
implementation) and are passed to the application entry point as well.

Removes some unecessary junk from main.cpp and makes it work with the
new oberon::application::run and platform changes.

Adds sequences to memory.hpp. These are mutable sequences equivalent to
a pointer to an array.