Add Vulkan device initialization.
Implements device selection and initialization for Vulkan. Devices are
selected manually with the 0th device being initialized by default. A
graphics and a presentation queue are also prepared (they are likely to
be the same queue in most cases).

Existing source files are updated to support this new functionality.
Also main.cpp is altered, as usual, to reflect so arbitrary testing I
did.