Add rendering implementation.
This is the beginnings of a rendering implementation. Most of the
allocation and selection of Vulkan assets works as intended but there
are still crucial bugs. Most importantly, attempting to resolve window
resizes while in the middle of a rendering step causes the application
to crash and causes Kwin to crash (on my system).

This also adds functionality to oberon::linux::window and
oberon::platform for tracking window resizes (and identifying them).