Hello triangle!
Implements a parameterless test frame shader program. This draws a
single triangle on the screen as a rendering configuration test/example.

To support the above, this adds a simple vertex and fragment shader
under data/oberon/shaders.

Infrastructure is also added to the oberon::graphics object for loading
and selecting shader programs (i.e., VkGraphicsPipelines). Currently,
I'm not sure this is a good long term solution. Shader modules and
programs are cleared whenever the device changes. This is necessary,
obviously, but it would be nice to have a narrow set of shaders that
could be rebuilt each time the device is swapped.

Existing infrastructure in oberon::graphics::begin_frame and
oberon::graphics::end_frame has been changed. Mostly, this is to clear
up validation warnings as well as a somewhat subtle bug involving image
memory barriers.

An oberon::graphics::draw method is added. This is probably going to
change drastically. The current implementation exists exclusively to
support the current test frame output.

Finally, oberon::system now has a method
(oberon::system::executable_path) that returns the directory under which
the program executable exists.