Refactor oberon::graphics.
Primarily, this removes the requirement (and ability) to load arbitrary
shaders using the graphics object. Instead, predefined shaders are
loaded using the system::find_file method. A number of API features are
also added to make it obvious what state the renderer is in (e.g.,
opened, closed, dirty). The most important change is the inclusion of
features to explicitly open and close graphics devices. This is now
required for proper operation. That is to say, after platform
initialization, client code is required to manually select one of the
available graphics devices. To simplify this, clients may call
graphics::preferred_device to select the implementation's preference.

To support these changes the oberon::linux::graphics implementation
is updated significantly. Most importantly, this fixes bugs related to
the dynamic rendering feature and Mesa graphics driver. Hopefully this
will prevent deadlocks and device losses in the future.

Finally, a to_string function is added for the pipeline_stage
enumeration.