Begin working on mesh allocation.
This adds several features. First, VulkanMemoryAllocator is added as a
dependency (only the header is required) as well as libglm. Second, a
VmaAllocator is now created (and later destroyed) during device
initialization. Finally, and most drastically, this changes the
architecture of the rendering process so that two primary command
buffers are submitted.

Basically, now there are two types of commands sent to the device queue.
First, there are transfer commands that move data to the device. Second,
there are rendering commands that do something with the data. Each frame
has a primary buffer for each and both are submitted at the same time.
Similarly, now each frame has its own command pool so that at the top of
the frame (after ensuring that it is no longer in flight) all associated
command buffers can be reset. This supresses a validation warning.

My thinking is that meshes (and other resources) would be supplied with
their own secondary command buffers that could be executed into the
primary buffers as need be.