Add vk_device object.
This is the first step to breaking out Vulkan device functionality into
its own object(s). The goal here is to allow the linux::graphics object
to simply allocate an appropriate vk_device rather than having to handle
all of the functionality directly.

For the most part, vk_device functionality is uniform across devices but
some functionality (e.g., queue selection) can be performed in a device
specific way by overriding the base methods. This is what the
*_vk_device classes do.

In the future, hopefully, this will form a secondary layer under
linux::graphics that allows handling Vulkan devices in a more succint
way.