Add result output to error handling.
Previously, when an error occurred, the application object would catch
any uncaught exception, print the associated message, and exit with a
non-zero status. This is fine but it's also useful to see the full
status code printed to stderr. For example, on Linux systems the full
error is truncated to 8-bits which might provide erroneus information if
it's not reported correctly otherwise.

Anyway, now the status is reported on error as well.