Complete cube example implementation.
This finishes up the requirements for the cube example. The example
program now draws a rotating 3D cube using the unlit vertex position and
color pipeline. The cube's rotation is independent of the frame rate
(i.e., it uses delta time) and the shader properly computes the image
using model, view, and projection matrices.