Decouple window and renderer.
This splits the render_window into two smaller more managable objects.
There are several major differences with this approach. In the old
approach a render_window would draw directly to swapchain images.
oberon::renderer will instead allocate and draw to its own images. These
images are then copied/blit (as appropriate) onto the swapchain. This
allows for rendering at a resolution separate from the swapchain
resolution. When blitting the image is upscaled using nearest-neighbor
scaling.

oberon::windows are only owned by the oberon::renderer while a frame is
live. As soon as all frames are submitted the relationship between the
two objects is broken. Hopefully, in the future this will allow reusing
the renderer object for off-screen images.