Infrastructure types and functions.
This adds handle types to types.hpp which are just infrastructure for
frame graphs. My intention is to reimplement the current renderer to use
a frame graph approach in the near future.

This also adds some template functions that can be used to find the
max/min size and alignment of a set of types. These are infrastructure
for a future transient object pool system.